Fixed lot of issues

+++ src/main/java/com/clm/matching/models/VendorDetailMatchResponseDTO.java
@@ -16,6 +16,6 @@ public class VendorDetailMatchResponseDTO {
     private Long id;
     private String name;
     private String description;
-    private double matchPercentage;
+    private int matchPercentage;
     List<CategoryMatchResponseDTO> categoryMatches;
 }
